+++ dash/dcc/Clipboard.py
@@ -4,13 +4,10 @@ import typing  # noqa: F401
 from typing_extensions import TypedDict, NotRequired, Literal  # noqa: F401
 from dash.development.base_component import Component, _explicitize_args
 
+ComponentSingleType = typing.Union[str, int, float, Component, None]
 ComponentType = typing.Union[
-    str,
-    int,
-    float,
-    Component,
-    None,
-    typing.Sequence[typing.Union[str, int, float, Component, None]],
+    ComponentSingleType,
+    typing.Sequence[ComponentSingleType],
 ]
 
 NumberType = typing.Union[
@@ -24,6 +21,10 @@ class Clipboard(Component):
 
     Keyword arguments:
 
+    - children (a list of or a singular dash component, string or number; optional):
+        Children rendered inside the Clipboard button before copying. By
+        default, a copy icon.
+
     - id (string; optional):
         The ID used to identify this component.
 
@@ -33,12 +34,16 @@ class Clipboard(Component):
     - content (string; optional):
         The text to be copied to the clipboard if the `target_id` is None.
 
+    - copied_children (a list of or a singular dash component, string or number; optional):
+        Children rendered inside the Clipboard button after the value has
+        been copied. By default, a check icon.
+
     - html_content (string; optional):
         The clipboard html text be copied to the clipboard if the
         `target_id` is None.
 
     - n_clicks (number; default 0):
-        The number of times copy button was clicked.
+        The number of times Clipboard button was clicked.
 
     - target_id (string | dict; optional):
         The id of target component containing text to copy to the
@@ -49,14 +54,16 @@ class Clipboard(Component):
     - title (string; optional):
         The text shown as a tooltip when hovering over the copy icon."""
 
-    _children_props: typing.List[str] = []
-    _base_nodes = ["children"]
+    _children_props: typing.List[str] = ["copied_children"]
+    _base_nodes = ["copied_children", "children"]
     _namespace = "dash_core_components"
     _type = "Clipboard"
 
     def __init__(
         self,
+        children: typing.Optional[ComponentType] = None,
         id: typing.Optional[typing.Union[str, dict]] = None,
+        copied_children: typing.Optional[ComponentType] = None,
         target_id: typing.Optional[typing.Union[str, dict]] = None,
         content: typing.Optional[str] = None,
         n_clicks: typing.Optional[NumberType] = None,
@@ -67,9 +74,11 @@ class Clipboard(Component):
         **kwargs
     ):
         self._prop_names = [
+            "children",
             "id",
             "className",
             "content",
+            "copied_children",
             "html_content",
             "n_clicks",
             "style",
@@ -78,9 +87,11 @@ class Clipboard(Component):
         ]
         self._valid_wildcard_attributes = []
         self.available_properties = [
+            "children",
             "id",
             "className",
             "content",
+            "copied_children",
             "html_content",
             "n_clicks",
             "style",
@@ -91,9 +102,9 @@ class Clipboard(Component):
         _explicit_args = kwargs.pop("_explicit_args")
         _locals = locals()
         _locals.update(kwargs)  # For wildcard attrs and excess named props
-        args = {k: _locals[k] for k in _explicit_args}
+        args = {k: _locals[k] for k in _explicit_args if k != "children"}
 
-        super(Clipboard, self).__init__(**args)
+        super(Clipboard, self).__init__(children=children, **args)
 
 
 setattr(Clipboard, "__init__", _explicitize_args(Clipboard.__init__))