+++ dash/dcc/Graph.py
@@ -7,13 +7,10 @@ from dash.development.base_component import Component, _explicitize_args
 from plotly.graph_objects import Figure
 
 
+ComponentSingleType = typing.Union[str, int, float, Component, None]
 ComponentType = typing.Union[
-    str,
-    int,
-    float,
-    Component,
-    None,
-    typing.Sequence[typing.Union[str, int, float, Component, None]],
+    ComponentSingleType,
+    typing.Sequence[ComponentSingleType],
 ]
 
 NumberType = typing.Union[